Allow class overriding
I wanted to experiment with a subtle behaviour change and stumbled upon a limitation that I could easily fix by extending my own Router. In theory it should be fine, but some methods and instance variables are set to `private`, when they could be set to `protected` and allow overriding.